Correct receive procedure for 3 bytes

+++ embedded_board/driver/serverCommunication.h
@@ -21,7 +21,7 @@ public:
 
 	serverCommunication();
 	~serverCommunication();
-	uint8_t readCredit(uint8_t ID);
+	uint32_t readCredit(uint8_t ID);
 	void writeCredit(uint8_t ID, uint8_t credit);
 };
